+++ app/policies/std_slot_policy.rb
@@ -1,22 +1,7 @@
 class StdSlotPolicy < SlotPolicy
-  # true if the user is signed in
-  def create_stdslot?
-    current_user?
-  end
-
   # TODO: write spec
   def slotsets?
     current_user_has_read_access?
-  end
-
-  # TODO: remove this
-  # true if the user is signed in and
-  # the user is slot owner
-  # meta data? can they only be updated by the creator?
-  def update_stdslot?
-    return false unless current_user?
-    return true if current_user == slot.owner
-    false
   end
 
   # true if the user is signed in and