+++ spec/policies/std_slot_policy_spec.rb
@@ -5,17 +5,7 @@ describe StdSlotPolicy do
 
   let(:slot) { create(:std_slot) }
 
-  permissions :create_stdslot? do
-    context "for a user" do
-      let(:user) { create(:user) }
-
-      it "allows access" do
-        expect(subject).to permit(user, slot)
-      end
-    end
-  end
-
-  permissions :update?, :update_stdslot?, :delete? do
+  permissions :update?, :delete? do
     let(:owner) { create(:user) }
     let(:creator) { create(:user) }
 
@@ -46,9 +36,7 @@ describe StdSlotPolicy do
   end
 
   describe 'for a visitor / invalid or missing auth_token' do
-    let(:permissions) {
-      [:create_stdslot?, :update?, :update_stdslot?]
-    }
+    let(:permissions) { [:update?] }
     let(:user) { nil }
 
     it "raises MissingCurrentUserError" do