+++ spec/routing/v1/slots_routing_spec.rb
@@ -21,33 +21,9 @@ RSpec.describe V1::SlotsController, type: :routing do
         .to route_to("v1/slots#delete", id: "1", format: :json)
     end
 
-    # TODO: remove this
-    it "routes to #create_stdslot" do
-      expect(post: "/v1/stdslot")
-        .to route_to("v1/slots#create_stdslot", format: :json)
-    end
-
-    # TODO: remove this
-    it "routes to #update_metaslot" do
-      expect(patch: "/v1/metaslot/1")
-        .to route_to("v1/slots#update_metaslot", id: "1", format: :json)
-    end
-
-    # TODO: remove this
-    it "routes to #update_stdslot" do
-      expect(patch: "/v1/stdslot/1")
-        .to route_to("v1/slots#update_stdslot", id: "1", format: :json)
-    end
-
     it "routes to #update" do
       expect(patch: "/v1/slots/1")
         .to route_to("v1/slots#update", id: "1", format: :json)
-    end
-
-    # TODO: remove this
-    it "routes to #destroy_stdslot" do
-      expect(delete: "/v1/stdslot/1")
-        .to route_to("v1/slots#delete", id: "1", format: :json)
     end
 
     it "routes to #slotsets" do
@@ -108,11 +84,6 @@ RSpec.describe V1::SlotsController, type: :routing do
     it "routes to #add_to_groups" do
       expect(post: "/v1/slots/1/slotgroups")
         .to route_to("v1/slots#add_to_groups", id: "1", format: :json)
-    end
-
-    it "routes to #reslot_history" do
-      expect(get: "/v1/slots/1/history")
-        .to route_to("v1/slots#reslot_history", id: "1", format: :json)
     end
 
     it "does not route to #new" do